feat(rail): move the tasks-rail toggle from the status bar to the top bars
The rail open/close control now lives at the right end of each surface's top bar, immediately after the primary 'New …' action, instead of the bottom status bar. A shared TaskRailToggle (top-bar icon styling, keeps the running-count dot + activity flash) drops into the viewer toolbar (covers every viewer + the markdown editor), the agents tab bar, the tasks board header, and the cabinet/room header. Removed from the status bar; ⌥⌘L still toggles it globally for the remaining utility surfaces.
