+++ src/components/agents/v2/tabs-layout.tsx
@@ -29,6 +29,7 @@ import { HeartbeatsTab } from "./heartbeats-tab";
 import { ScheduleView } from "@/components/cabinets/schedule-view";
 import { ContentSheet } from "@/components/layout/content-sheet";
 import { FolderTabs } from "@/components/layout/folder-tabs";
+import { TaskRailToggle } from "@/components/tasks/rail/task-rail-toggle";
 
 export type AgentsTabKey = "agents" | "routines" | "heartbeats" | "schedule" | "channels";
 
@@ -157,6 +158,7 @@ function TopBar({
         <Divider className="hidden md:block" />
         <MasterToggle />
         <NewButton tab={tab} />
+        <TaskRailToggle />
       </div>
     </header>
   );
